+++ vyxal/elements.py
@@ -5703,7 +5703,7 @@ def vy_exec(lhs, ctx):
         stack = ctx.stacks[-1]
         exec(
             vyxal.transpile.transpile(
-                lhs, ctx.dictionary_compression, ctx.variable_length_1
+                lhs, options=ctx.transpilation_options
             )
         )
