+++ vyxal/main.py
@@ -9,10 +9,10 @@ import traceback
 import types
 
 import vyxal.encoding
-from vyxal.context import Context
+from vyxal.context import Context, TranspilationOptions
 from vyxal.elements import *
 from vyxal.helpers import *
-from vyxal.transpile import transpile, TranspilationOptions
+from vyxal.transpile import transpile
 
 THIS_FOLDER = os.path.dirname(os.path.abspath(__file__)) + "/.."
 sys.path.insert(1, THIS_FOLDER)
@@ -162,6 +162,12 @@ def execute_vyxal(file_name, flags, inputs, output_var=None, online_mode=False):
     ctx.double_zip_vectorize = "Z" in flags
     ctx.utf8strings = "U" in flags
 
+    options = TranspilationOptions()
+    options.dict_compress = ctx.dictionary_compression
+    options.variables_as_digraphs = ctx.variable_length_1
+    options.utf8strings = ctx.utf8strings
+    ctx.transpilation_options = options
+
     if "2" in flags:
         ctx.default_arity = 2
     elif "3" in flags:
@@ -170,10 +176,6 @@ def execute_vyxal(file_name, flags, inputs, output_var=None, online_mode=False):
         ctx.default_arity = 1
 
     try:
-        options = TranspilationOptions()
-        options.dict_compress = ctx.dictionary_compression
-        options.variables_as_digraphs = ctx.variable_length_1
-        options.utf8strings = ctx.utf8strings
         code = transpile(code, options)
     except Exception as e:  # skipcq: PYL-W0703
         if ctx.online:
